# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API MiniMarket Plus",
        "description": "Microservicio backend de MiniMarket Plus. Expone la gestion de\nproductos, categorias, inventario, carrito de compras, ventas,\ndetalle de ventas y usuarios del sistema.\n\nEsta documentacion se genera con OpenAPI Specification (OAS 3)\nmediante springdoc-openapi, asegurando la consistencia entre el\ncodigo y la documentacion de la API.",
        "contact": {
            "name": "Equipo Backend - MiniMarket Plus",
            "email": "soporte@minimarketplus.cl"
        },
        "license": {
            "name": "Uso academico Duoc UC",
            "url": "https://www.duoc.cl"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Entorno de desarrollo local"
        }
    ],
    "security": [
        {
            "autenticacionBasica": []
        }
    ],
...
    "paths": {
...
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Obtener categoria por ID",
                "description": "Devuelve la informacion detallada de una categoria segun su identificador.",
                "operationId": "obtenerCategoriaPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la categoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "200": {
                        "description": "Categoria encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Actualizar categoria",
                "description": "Actualiza los datos de una categoria existente identificada por su ID.",
                "operationId": "actualizarCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la categoria a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "200": {
                        "description": "Categoria actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Eliminar categoria",
                "description": "Elimina una categoria del catalogo por su identificador.",
                "operationId": "eliminarCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la categoria a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoria eliminada (sin contenido)"
                    },
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                }
            }
        },
...
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias",
                "description": "Obtiene la lista completa de categorias que agrupan los productos de MiniMarket Plus.",
                "operationId": "listarCategorias",
                "responses": {
                    "401": {
                        "description": "No autenticado"
                    },
                    "200": {
                        "description": "Lista de categorias obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Crear categoria",
                "description": "Registra una nueva categoria para agrupar los productos del catalogo de MiniMarket Plus.",
                "operationId": "guardarCategoria",
                "requestBody": {
                    "description": "Datos de la categoria a crear",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            },
                            "examples": {
                                "Categoria nueva": {
                                    "description": "Categoria nueva",
                                    "value": {
                                        "nombre": "Abarrotes"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "No autenticado"
                    },
                    "200": {
                        "description": "Categoria creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categoria": {
                "required": [
                    "nombre"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico de la categoria.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre unico de la categoria.",
                        "example": "Bebidas"
                    },
                    "productos": {
                        "type": "array",
                        "description": "Productos asociados a esta categoria.",
                        "readOnly": true,
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "description": "Categoria que agrupa productos del catalogo (por ejemplo Bebidas o Abarrotes)."
            },
            "Producto": {
                "required": [
                    "categoria",
                    "nombre",
                    "precio",
                    "stock"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Identificador unico del producto.",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "description": "Nombre comercial del producto.",
                        "example": "Agua Mineral 1L"
                    },
                    "precio": {
                        "type": "number",
                        "description": "Precio de venta unitario en pesos chilenos.",
                        "format": "double",
                        "example": 890.0
                    },
                    "stock": {
                        "type": "integer",
                        "description": "Cantidad disponible en stock.",
                        "format": "int32",
                        "example": 50
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                },
                "description": "Producto disponible en el catalogo de MiniMarket Plus."
            },
...
        },
        "securitySchemes": {
            "autenticacionBasica": {
                "type": "http",
                "description": "Autenticacion HTTP Basic. Use el usuario y la contrasena semilla (admin / admin123) en el boton Authorize para consumir los endpoints protegidos.",
                "name": "autenticacionBasica",
                "scheme": "basic"
            }
        }
    }
}